shifted location of conversion patterns from MappingStrategyBuilder to ConversionPatternRepository
This should make the integration with IoC containers much easier, since now only ConversionPatternRepository knows and controls how conversion patterns get registered and built.
removed IRootConversionPattern. It was a sucky abstraction (almost like a leaky abstraction :) and it was really a hack. Now the call to map is handled explicitly in code.